Reverted all changes that were causing a performance degradation
